添加 serialization, displayObject, inherit 配置. 更新 displayPlace 配置

+++ src/main/java/aplini/usetranslatednames/Enum/Status.java
@@ -3,5 +3,5 @@ package aplini.usetranslatednames.Enum;
 public class Status {
     public long Messages = 0;
     public long Matches = 0;
-    public long TotalTime = 0;
+    public double TotalTime = 0;
 }
